Fix some problems with icon paths when configured to a prefix other than /usr.
Some fixes for the kernel driver (although I cannot yet get it working). Make sure all dependencies are checked in configure.in (discovered some missing checks while trying to build on openSuSE). 

Also made core plugins optional if dependencies are not there. This is a temporary measure until the re-arrangement of the plugin structure is complete.