Do not hang at first launch after install
When checking and enabling new plugins, the instance of G15Plugins does not have
a 'screen' attached to it.
But in the _plugin_changed method that is called when the plugin is enabled a
call is made to a 'screen' instance.
A AttributeError: 'NoneType' object has no attribute '_check_active_plugins'
is thrown and g15-desktop-service just hanged in there and a KILL signal had to
be sent.

Fixes #167