weather: display error message if error parsing weather data
If there is an error parsing the weather data fetched either by Yahoo or
NOAA, the plugin displayed the raw SVG.

While the parsing errors should be fixed in the data sources themselves, it is
a good idea to have the weather plugin to display a generic error message.

This way, even if a 3rd party data source bugs, the weather plugin will have a
acceptable behavior.

refs #183