g15-config-tool: display a error message when importing macro profile for wrong device
When importing a macro profile from a mzip file, there was no control on which
device the profile applied to.
The macro profile was created, but then didn't appear on the profile list.
Now, if the user tries to import a macro profile for a device other than the
selected one, a error message is displayed, and the profile is not created.

refs #192