g15-config-tool: do not copy extra files if incompatible macro profile
When importing a macro profile made for another device, there was a chance that
some extra files from that profile were copied (e.g. icons files).
This happened because the parsing of the .macros file was not done at first
place so that we could check the device for which the profile was made.

fixes #192