mpris-plugin: Fix problems with players not 100% compliant with D-Bus specification
Some players (at least vlc) do not seem to be fully compliant with the D-Bus
specification [1].
In the case of vlc, it does not implement the org.freedesktop.DBus.Properties.GetAll
method.
This method was used in the mpris plugin to get some information about the
player and the currently playing track.

Since the properties which are needed are know, another method is now used to
fetch their values.

[1] http://dbus.freedesktop.org/doc/dbus-specification.html

fixes #187