voip-teamspeak3: take subchannels into account when parsing channelconnectinfo response
When a user goes into a subchannel in teamspeak3, the channelconnectinfo response
contains the full path of the channel separated by forward slashes.

When the plugin tries to find the current channel of a user, it tried to match
the full path against only the channel name. As a result, it never found the
channel and returned None.
This made the calling code throw the AttributeError, since it tried to get the
name field from a 'None' object.

By storing the full path of a channel when this one is created and by doing the
match against it, the channel is now found and the AttributeError is no longer
thrown

refs #188