voip-teamspeak3: take new channel_order value into account when editing a channel
In TeamSpeak3 UI, a user can update the a channel's order in the "Edit Channel"
window.
The new value for the order is now taken into account, and the edited channel
order is updated, and the channel is placed into it's correct place in the
channels list.
The position of children of the updated channel are also updated.

refs #188