voip-teamspeak3: also delete subchannels on channel removal
Teamspeak3 sends a 'notifychanneldeleted' message with the list of cids of all
the removed channels when a channel is removed from the server.
The plugin code was usually expecting a single cid and an error was thrown.
Now, the code takes into account the fact that there maybe several cids passed
and always use the latest cid, since it is always the one of the 'parent'
channel that is removed.

refs #188