weather: Change the default refresh time to 60 minutes
Gnome15's weather plugin default refresh interval was not being handled
correctly.
By default it's value was being set to 3600 minutes, but on the settings window
the displayed value was always "1", until the user manually modified this value
at least once.

The default value is now set to 60 minutes, and this value is now correctly
displayed.

fixes #174