extract scheduler code from g15util to a new file
This is the beginning of a operation aiming to split the g15util.py file which
has become a dump bin for a lot of misc things.

The goal of this split is to allow g15-system-service to automatically start on
Fedora. Currently, we have a "Could not open display" runtime error.

When g15util.py is loaded, it tries to import gtk.gdk. The gtk module tries to
access a running X instance when loaded.
Unfortunately, the dbus environment isn't aware of any running instance and the
runtime error above is thrown.
Since g15-system-service doesn't depend on gtk, after taking out all the code
depending on gtk to other files, this problem should be solved.

refs #173