g15-config: do not instantiate G15GlobalPlugins several times
Currently, each time the user clicked on the "Configure" button of the
configuration window, a new instance of G15GlobalConfig was created.
As a consequence, since the widget signals are connected in the constructor,
some methods were called multiple times (once per G15GlobalConfig
instantiation).

For instance, if the user clicked the "Configure" button three times during the
execution of g15-config, if the user clicked the "Preferences" button for a
global plugin, the plugin preferences window would be shown three times.

We now only create one instance of G15GlobalPlugins and reuse that instance each
time the user clicks the "Configure" button.

fixes #229