processes: correctly sets window icon in the applications list
The method used for getting the application icon in the processes plugin when
using wnck was trying to use the window.get_icon_name method of WnckWindow.

The get_icon_name method however returns the caption that is usually displayed
in a pager when the window is minimized and not the name of the themed icon used
for the window as in Gtk.Window.

fixes #234