g15-desktop-service: detach the G19 USB interfaces when the keyboard is disconnected
Previously, when the G19 driver was "disconnected", pylibg19 kept the USB
interfaces opened.
This had as a side effect that if g15-desktop-service was "re-enabled"
(e.g after switching to a VT console and returning to X), g15-desktop-service
wasn't able to reconnect to the keyboard.

We now explicitly detach the USB interfaces, so that they can be re-attached
when the driver is re-connected.

closes #208