voip-teamspeak3: Handle cases where the client_unique_identifier is not base64 encoded.
In some cases, the client_unique_identifier is not base64 encoded.
It seems to be the case for "special" users, like the serveradmin and
ServerQuery.

When this happened, a TypeError was thrown. We now fallback to use the non
encoded value, if the client_unique_identifier cannot be base64 decoded.

fixes #233