tweak: increase the maximum value of the key hold duration to 10[s]
The KeyHoldDuration setting adjustement in the Tweak plugin preferences dialog
was set to 100.
The default value for this setting is set to 2000.

Each time the Tweak plugin preferences dialog was displayed, the maximum value
was erroneously set to 100, which is wrong.

The maximum value for this setting is now 10000 (10[s]).

fixes #227