g15-systemtray: enable monitoring of Gnome15 gconf directory
g15-systemtray tried to react to a change of state of the "Cycle Screens"
setting and depending on the value of that setting, the corresponding menu item
was checked or not.

However, to be informed of a change on a given gconf key, gconf must be told to
monitor a gconf directory above the key to be monitored.

This wasn't done and as such, g15-systemtray was never informed of the changes
of the "Cycle Screens" key and as such didn't react to those changes.

fixes #246