move gnome15-gaming-plugins into gnome15 folder
This subproject had only a plugin (game-nexuiz).
This plugin is now on the same level than the other gnome15 plugins.
By default it is not yet installed, since it seem to be a work in progress
task that the previous maintainer had.

The autotools files were adapted for at least to install this plugin correctly.

refs #171