organize dependencies on configure.ac file
Currently, the configure.ac file is a bit desorganized.

The next commits will try to organize that file in this order:

- autotools initialization
- check for dependencies
- parse command line arguments
- create output files
- display configuration results

Some changes on the name of the arguments and variables used will also
be made to keep the file consistent.

This commit, puts all the checks for dependencies after the autotools
initialization macros.

refs #171