build: use environment variables for some configure parameters
The AC_ARG_ENABLE and AC_ARG_WITH macros seems to be dedicated
to enable a certain feature or to specify external dependencies
of a project.

Locales to build, fixed size font name, udev rules path, and devices
permissions do not belong to those categories.

The usage of AC_ARG_VAR seems to be more appropriate in this case.

refs #171