build: use consistent parameters for the drivers
From now on, to enable or disable a driver, configure must be passed
the --enable-driver-<driver_name> or --disable-driver-<driver_name>.

By default the drivers are built if their dependencies are installed.

If a user enables the build of a driver and it's dependencies are not
installed, an error message is show, and the configure process is
aborted.

refs #171