build: use consistent parameters for the plugins
From now on, to enable or disable a plugin, configure must be passed
the --enable-plugin-<driver_name> or --disable-plugin-<driver_name>.

By default, except for the experimental plugins, all the plugins are
built if their dependencies are installed.

If a user enables the build of a plugin and it's dependencies are not
installed, an error message is show, and the configure process is
aborted.

refs #171