build: correctly set the default value to enable the indicator
When building Gnome15, by default it will try to build the
indicator, even if appindicator is not installed.
This causes an error, and the user must pass the --disable-indicator
argument to ./configure.

This commit fix this problem by only enabling build of the indicator
if the appindicator python module is installed.

fixes #258