screensaver: handle connection errors to the screensaver dbus interface.
To connect to the screensaver dbus interface, gnome15 tries a number
of different combinations between the dbus_name, dbus_interface and
object to support the different desktop environments.
It used to consider that he had found a valid implementation
if it could create a proxy.
However, dbus-python allows you to create a proxy to a non existent
object.

For instance, it could create a interface with
dbus_name=org.gnome.ScreenSaver, dbus_interface=org.gnome.ScreenSaver,
object_path=/, event if there was no object implementing the
org.gnome.ScreenSaver at the object_path /.
This made the screensaver plugin throw a DBUSException when trying to
check if the screensaver was active.

Now, the plugin checks that it can effectively call methods on the
created proxy before consider that he found a valid interface.

refs #257