sense: enables udisks2 as a sensor source
The introduction of udisks2 as a possible sensor source raised a
problem.
There now may be several sensor sources providing the same information
to the sense plugin.
For instance, since a system can have both udisks and udisks2 installed
there would be two sources providing hard drive temperature information
to sense.

This patch allows tackles that problem by grouping the sensor sources
that provide the same data.
When checking for the sensors sources available, the sense plugin
checks the sensor sources of a given group until it finds a working
one. It then skips the other sensors sources of the same group.

closes #261