sense: doesn't read udisks sensor source values only once
The UDisksSource._check_dbus_connection method that was being used to
both get the list of available sensors and to get the values from those
sensors.

However, this method had a condition that made it only been executed
once and as such, the plugin keep displaying the same values.

The logic of this method splitted between the is_valid and the get_sensors
method.

fixes #262