libimpulse: link missing libraries
Changes on commit 7f0709496d31d75c7383edf983fa78af8d9c23bc made that
libpulse and libfftw were no longer linked to libimpulse.

This caused a undefined symbol error when trying to load libimpulse and
as a consequence, the impulse15 plugin no longer worked.

We now, instead of checking for the libraries individually, use
pkg-config to init the _CFLAGS and _LIBS variables for those libraries
and link to them.

We now also use AC_CHECK_LIB instead of AC_SEARCH_LIBS for checking if
libg15 is available to ensure that libimpulse does not link to libg15.

fixes #267