util.g15uigconf: return the handler id for the created signals
The helpers provided by g15uigconf connect a handler function to some
signals emmited by the widget.

For instance, in configure_combo_from_gconf, the 'changed' signal is
connected to combo_box_changed, so that when the user selects a item
in the combobox, the associated gconf key is updated with the new
value.

Up until now, the handler id's were 'lost' and there way no way to
manipulate them again.

Sometimes however, client code may need to have access it, for
instance to block it temporarily.

The handler id is now returned.
Helpers that were already returning something, now return a tuple with
the handle id being it's first element.

refs #212