volume: ensure that configuration changes are only handled once
The _config_changed method was called twice if the user changed the
soundcard on the preferences dialog.
A first time for the soundcard change, and a second time because the
first mixer of the newly selected soundcard is automatically selected.
The volume monitoring would then be restarted twice, which makes no
sense.
Instead of restarting the monitoring as soon as this method is called,
we put it as a task on a queue for 1 second. If during that time, any
other change happens to the configuration, the previous restart request
is cancelled, and another one takes it's place.
This way, the monitoring is only restarted once when the user selects another
sound card.

fixes #212