creates virtual joysticks with their axis centered
Gnome15 virtual joysticks axis were setup to sent values between 0 for
the top and left positions and 255 for the right and bottom positions.
The center should then be at 128.

After creating the joystick virtual device, we sent a uinput command to
place the joystick axis at their center (position 128).

However, joysticks created by uinput seem to have a odd behaviour.
By default, when a virtual joystick device file (/dev/input/jsX) is
opened, it provides the list of axis and buttons with their current
state.
In the case of uinput created joysticks, it seems that their state is
not memorized between "open" operations.
The consequence of this, is that by default, all the buttons are
unpressed and the axis are at their 0,0 position, which means for
our joyticks, that their are at their top left position.

This causes problems in games that take joystick input into account.

We workaround this problem by changing the range of the joystick device
to be from -127 to 127 instead of 0 to 255.
This way, the center position is at 0, which means that when programs
open the joystick file (/dev/input/jsX) the position that it supplies
by default corresponds to the center and not to the top-left.

refs #271