driver-kernel: fix digital-joystick mode for G13 devices
uinput codes are stored as two value tuples.
The first value corresponds to the type of the event.
The second value corresponds to the code of the event.

When emulating a digital joystick the kernel driver was emitting values
for events of type EV_REL (relative events).
Joysticks emit EV_ABS events (absolute events).

fixes #275