g15-config: remove the "Driver" tab
UI components are moved to the "Keyboard" tab.
The options of the driver are now displayed when the user click the new
Options button.

If no driver is available for a given device, the tabs are no longer
displayed and a message is displayed to the user instead.

While these changes do simplify the UI, I made them mainly because
I couldn't solve issue #276 differently.
The driver options were being displayed directly in the driver tab
itself, but this seemed to cause a problem with the kernel driver
options where there was an attribute error.

fixes #276