weather: update data also even if not displayed
The weather plugin was set up to only refresh itself when its page
was displayed on the device screen.

Operating this way, the user had to have the plugin displayed at least
the value of its update setting before the data was refreshed.
With a "normal" use of the plugin, the user could have the impression
that the data was never updated.

refs #298