g15plugin: fix behavihor when only_refresh_when_visible is False
When the value of G15RefreshingPlugin.only_refresh_when_visible was
False, the _schedule_refresh method was never called.
This meant that the plugin data was never refreshed unless it was
asked to do so explicitely.

This change forces the _shedule_refresh method to be called at the
plugin activation.
The _schedule_refresh method now always schedules a call to the
_refresh method, which will ensure that the plugin data is refreshed if
needed.

refs #298