g19: avoid excessive logging
pylibg19 continually tries to read the values of the G19 special keys.
When the G19 has no data, a "USBError: Connection timed out" exception
is raised.
Previously, we were logging every one of this errors which consumes
resources and does not provide any benefit to debugging.

We now skip this precise error, however if other kinds of USBErrors
are raised, they are logged.

refs #269